agents/openai-retrieval-agent-dabs/requirements.txt
mlflow>=2.21.0
databricks-sdk>=0.40.0
databricks-agents>=0.16.0
databricks-vectorsearch>=0.40.0
openai>=1.50.0
pydantic>=2.0.0
backoff>=2.0.0
PyYAML>=6.0
